refactor: Change search and filter methods to try/catch way
In this version of the function the route handler wraps the calls to 'searchPatientsByDeviceId'
and 'filterPatientsAlreadyAddedByAdminEmail' in a try/catch block.

If either of these functions throw an error, the catch block
will catch the error and execute the failed function, passing the
error message as an argument.

If the try block executes successfully, the successed function is called with the data returned by the functions.